stm32: Support LPTIM v1 and v2
